Add CreateProductDto tests + fixes
Changes Product price field from a double to a decimal for precise calculations

+++ product-catalog-service/ProductAPI/DTOs/GetProductDto.cs
@@ -5,6 +5,6 @@ public class GetProductDto
     public required int Id { get; set; }
     public required string Name { get; set; }
     public required string Description { get; set; }
-    public required double Price { get; set; }
+    public required decimal Price { get; set; }
     public required int Stock { get; set; }
 }
